Added wheel drawing also
At current the car needs to be registered with a picture larger than it's physical body in order to get the front wheels in the current position for steering. This is a little unintuitive. It may be worth adding some additional parameters to control axel positions though the number of params in the car class is already rather high. May require a rethink, especially given the image ids now creeping in.